Latest changes to ticket #698 and ticket #451

+++ vede/app/controller/DeviceEditor/InspectorController.js
@@ -58,10 +58,25 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
     },
 
     checkCombinatorial:function(j5collection,cb){
+        var tmpC = 0;
         combinatorial = false;
         j5collection.bins().each(function(bin,binKey){
-            if(bin.parts().getCount()>1) combinatorial = true;
+            if(bin.parts().getCount()>1) {
+                bin.parts().each(function(part) {
+                    part.getSequenceFile({
+                        callback: function(sequenceFile){
+                        if(sequenceFile.get("partSource")!="") {
+                            tmpC++;
+                        }
+                    }
+                });
+            });
+        }
         });
+        console.log(tmpC);
+        if (tmpC>0) {
+            combinatorial = true;
+        }
         return cb(combinatorial);
     },
 
@@ -84,6 +99,7 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
                 }
                 else {j5ready = false;}
             });
+            console.log(combinatorial);
             tab.query("component[cls='combinatorial_field']")[0].setValue(combinatorial);
             tab.query("component[cls='j5_ready_field']")[0].setValue(j5ready);
 
@@ -155,9 +171,6 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
             // If the part is not owned by a bin yet, add it to the bin.
             if(this.DeviceDesignManager.getBinAssignment(this.activeProject,
                                                          this.selectedPart) < 0) {
-                var added = this.DeviceDesignManager.addPartToBin(this.activeProject,
-                                                      this.selectedPart,
-                                                      this.selectedBinIndex);
                 var selectedBinIndex = this.selectedBinIndex;
             }
 
@@ -225,8 +238,10 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
                                         callback: function(sequence){
                                             if(bin)
                                             {
+                                                console.log(self.selectedPart);
                                                 var insertIndex = bin.parts().indexOf(self.selectedPart);
                                                 // var binIndex = self.DeviceDesignManager.getBinIndex(self.activeProject,bin);
+                                                console.log(insertIndex);
                                                 bin.parts().removeAt(insertIndex);
                                                 bin.parts().insert(insertIndex,part);
                                                 self.onReRenderDECanvasEvent();
@@ -280,6 +295,9 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
 
         openPartLibraryBtn.enable();
         openPartLibraryBtn.removeCls('btnDisabled');
+
+        var removeRowMenuItem = this.tabPanel.down("button[cls='editMenu'] > menu > menuitem[text='Remove Row']");
+        removeRowMenuItem.enable();
 
         if(this.selectedBinIndex !== 0) {
             // Turn the FAS_LIST array into an array of arrays, as required by
@@ -458,7 +476,6 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
 
         if(selectedBin) {
             var selectedBinIndex = this.DeviceDesignManager.getBinIndex(this.activeProject, selectedBin);
-
             this.DeviceDesignManager.addEmptyBinByIndex(this.activeProject, (selectedBinIndex+1));
             this.columnsGrid.getSelectionModel().deselectAll();
         } else {
@@ -471,18 +488,36 @@ Ext.define("Vede.controller.DeviceEditor.InspectorController", {
      * is selected, removes the last bin in the design.
      */
     onRemoveColumnButtonClick: function () {
+
         var selectedBin = this.columnsGrid.getSelectionModel().getSelection()[0];
         var removeColumnMenuItem = this.tabPanel.down("button[cls='editMenu'] > menu > menuitem[text='Remove Column']");
-        
-        if(selectedBin) {
-            var selectedBinIndex = this.DeviceDesignManager.getBinIndex(this.activeProject, selectedBin);
-            this.activeProject.getJ5Collection().deleteBinByIndex(selectedBinIndex);
+
+        if (selectedBin) {
+            Ext.Msg.show({
+                    title: "Are you sure you want to delete this row?",
+                    msg: "WARNING: This will delete the current selected row. This process cannot be undone.",
+                    buttons: Ext.Msg.OKCANCEL,
+                    cls: "messageBox",
+                    fn: this.removeColumn.bind(this, selectedBin),
+                    icon: Ext.Msg.QUESTION
+            });
         } else {
-            this.activeProject.getJ5Collection().deleteBinByIndex(
-            this.activeProject.getJ5Collection().binCount() - 1);
+            removeColumnMenuItem.disable();
         }
 
         removeColumnMenuItem.disable();
+    },
+
+    removeColumn: function (selectedBin, evt) {
+        if (evt === "ok") {
+            if(selectedBin) {
+                var selectedBinIndex = this.DeviceDesignManager.getBinIndex(this.activeProject, selectedBin);
+                this.activeProject.getJ5Collection().deleteBinByIndex(selectedBinIndex);
+            } else {
+                this.activeProject.getJ5Collection().deleteBinByIndex(
+                this.activeProject.getJ5Collection().binCount() - 1);
+            }
+        }
     },
 
     reconfigureEugeneRules: function() {
